move action object to hxactions
